+++ DrawingApp/DrawingApp/Id.swift
@@ -1,0 +1,19 @@
+//
+//  Id.swift
+//  DrawingApp
+//
+//  Created by 김하림 on 2023/04/11.
+//
+
+import Foundation
+
+class Id {
+    private var id:String
+    
+    init(id: String) {
+        self.id = id
+    }
+    func getId() -> String {
+        return self.id
+    }
+}
