+++ DrawingApp/DrawingApp/RectangleSize.swift
@@ -1,0 +1,20 @@
+//
+//  RectangleSize.swift
+//  DrawingApp
+//
+//  Created by 김하림 on 2023/04/11.
+//
+
+import Foundation
+
+class RectangleSize {
+    private var width:Double = 150
+    private var height:Double = 120
+    
+    func getWidth() -> Double {
+        return self.width
+    }
+    func getHeight() -> Double {
+        return self.height
+    }
+}
